# Reduce time taken to collect VerifiedTransactions