--- a/zilliqa/src/consensus.rs
+++ b/zilliqa/src/consensus.rs
@@ -2617,6 +2617,8 @@ impl Consensus {
         transactions: Vec<SignedTransaction>,
         committee: &[NodePublicKeyRaw],
     ) -> Result<()> {
+        let len = transactions.len();
+        let start = std::time::Instant::now();
         debug!("Executing block: {:?}", block.header.hash);
 
         let parent = self
@@ -2627,8 +2629,10 @@ impl Consensus {
             trace!("applying {} transactions to state", transactions.len());
         }
 
+        let now = std::time::Instant::now();
         let transactions: Result<Vec<_>> = transactions.into_iter().map(|tx| tx.verify()).collect();
         let mut transactions = transactions?;
+        info!("{} ELAPSED_VER {:?}", len, now.elapsed());
 
         let transaction_hashes = transactions
             .iter()
@@ -2638,6 +2642,7 @@ impl Consensus {
         // We re-inject any missing Intershard transactions (or really, any missing
         // transactions) from our mempool. If any txs are unavailable either in the
         // message or locally, the proposal cannot be applied
+        let now = std::time::Instant::now();
         for (idx, tx_hash) in block.transactions.iter().enumerate() {
             if transactions.get(idx).is_some_and(|tx| tx.hash == *tx_hash) {
                 // all good
@@ -2649,6 +2654,7 @@ impl Consensus {
                 transactions.insert(idx, local_tx);
             }
         }
+        info!("{} ELAPSED_INS {:?}", len, now.elapsed());
 
         self.apply_rewards(committee, &parent, block.view(), &block.header.qc.cosigned)?;
 
@@ -2666,6 +2672,7 @@ impl Consensus {
         let mut receipts_trie = eth_trie::EthTrie::new(Arc::new(MemoryDB::new(true)));
         let mut transactions_trie = eth_trie::EthTrie::new(Arc::new(MemoryDB::new(true)));
 
+        let now = std::time::Instant::now();
         for (tx_index, txn) in transactions.into_iter().enumerate() {
             self.new_transaction(txn.clone())?;
             let tx_hash = txn.hash;
@@ -2702,6 +2709,7 @@ impl Consensus {
             info!(?receipt, "applied transaction {:?}", receipt);
             block_receipts.push((receipt, tx_index));
         }
+        info!("{} ELAPSED_APP {:?}", len, now.elapsed());
 
         if cumulative_gas_used != block.gas_used() {
             warn!("Cumulative gas used by executing all transactions: {cumulative_gas_used} is different that the one provided in the block: {}", block.gas_used());
@@ -2741,6 +2749,7 @@ impl Consensus {
             ));
         }
 
+        let now = std::time::Instant::now();
         for (receipt, tx_index) in &mut block_receipts {
             receipt.block_hash = block.hash();
             // Avoid cloning the receipt if there are no subscriptions to send it to.
@@ -2748,6 +2757,7 @@ impl Consensus {
                 let _ = self.receipts.send((receipt.clone(), *tx_index));
             }
         }
+        info!("{} ELAPSED_RCT {:?}", len, now.elapsed());
 
         // Important - only add blocks we are going to execute because they can potentially
         // overwrite the mapping of block height to block, which there should only be one of.
@@ -2761,16 +2771,18 @@ impl Consensus {
             // closure has to be move to take ownership of block_receipts
             let db = &self.db;
             self.db.with_sqlite_tx(move |sqlite_tx| {
+                let now = std::time::Instant::now();
                 for (receipt, _) in block_receipts {
                     db.insert_transaction_receipt_with_db_tx(sqlite_tx, receipt)?;
                 }
+                info!("{} ELAPSED_SQL {:?}", len, now.elapsed());
                 Ok(())
             })?;
         }
 
         // Tell the block store to request more blocks if it can.
         self.block_store.request_missing_blocks()?;
-
+        info!("{} ELAPSED_EXE {:?}", len, start.elapsed());
         Ok(())
     }